+++ packages/adapters/src/adapters/Cornerstone/Segmentation_4X.js
@@ -417,7 +417,7 @@ async function generateToolState(
     const labelmapBufferArray = [];
     labelmapBufferArray[0] = new ArrayBuffer(arrayBufferLength);
 
-    // Precompute the indices and metadata so that we don't have to call
+    // Pre-compute the indices and metadata so that we don't have to call
     // a function for each imageId in the for loop.
     const imageIdMaps = imageIds.reduce(
         (acc, curr, index) => {
@@ -434,7 +434,7 @@ async function generateToolState(
     // segment in the labelmapBuffer
     const segmentsPixelIndices = new Map();
 
-    await insertFunction(
+    const overlappingSegments = await insertFunction(
         segmentsOnFrame,
         segmentsOnFrameArray,
         labelmapBufferArray,
@@ -473,7 +473,8 @@ async function generateToolState(
         segMetadata,
         segmentsOnFrame,
         segmentsOnFrameArray,
-        centroids: centroidXYZ
+        centroids: centroidXYZ,
+        overlappingSegments
     };
 }
 
@@ -1141,6 +1142,7 @@ function insertPixelDataPlanar(
 
     const shouldTriggerEvent = triggerEvent && eventTarget;
 
+    let overlapping = false;
     // Below, we chunk the processing of the frames to avoid blocking the main thread
     // if the segmentation is large. We also use a promise to allow the caller to
     // wait for the processing to finish.
@@ -1244,6 +1246,9 @@ function insertPixelDataPlanar(
                     if (data[j]) {
                         for (let x = j; x < len; ++x) {
                             if (data[x]) {
+                                if (!overlapping && labelmap2DView[x] !== 0) {
+                                    overlapping = true;
+                                }
                                 labelmap2DView[x] = segmentIndex;
                                 indexCache.push(x);
                             }
@@ -1278,7 +1283,7 @@ function insertPixelDataPlanar(
                 setTimeout(processInChunks, 0);
             } else {
                 // resolve the Promise when all chunks have been processed
-                resolve();
+                resolve(overlapping);
             }
         }