+++ packages/core/src/RenderingEngine/StackViewport.ts
@@ -35,7 +35,7 @@ import type {
   VOIRange,
   VolumeActor,
 } from '../types';
-import { ViewportInput } from '../types/IViewport';
+import { TargetSpecifier, ViewportInput } from '../types/IViewport';
 import {
   actorIsA,
   colormap as colormapUtils,
@@ -2279,13 +2279,15 @@ class StackViewport extends Viewport implements IStackViewport, IImagesLoader {
 
     // 3b. If we cannot reuse the vtkImageData object (either the first render
     // or the size has changed), create a new one
+
+    const pixelArray = image.getPixelData();
     this._createVTKImageData({
       origin,
       direction,
       dimensions,
       spacing,
       numComps,
-      pixelArray: image.getPixelData(),
+      pixelArray,
     });
 
     // Set the scalar data of the vtkImageData object from the Cornerstone
@@ -2809,6 +2811,11 @@ class StackViewport extends Viewport implements IStackViewport, IImagesLoader {
   public getCurrentImageIdIndex = (): number => {
     return this.currentImageIdIndex;
   };
+
+  public getTargetId(specifier: TargetSpecifier = {}): string {
+    const { sliceIndex: imageIdIndex = this.currentImageIdIndex } = specifier;
+    return `imageId:${this.imageIds[imageIdIndex]}`;
+  }
 
   /**
    *