+++ packages/core/src/RenderingEngine/Viewport.ts
@@ -30,7 +30,11 @@ import type {
   EventTypes,
   DisplayArea,
 } from '../types';
-import type { ViewportInput, IViewport } from '../types/IViewport';
+import type {
+  ViewportInput,
+  IViewport,
+  TargetSpecifier,
+} from '../types/IViewport';
 import type { vtkSlabCamera } from './vtkClasses/vtkSlabCamera';
 import { getConfiguration } from '../init';
 import IImageCalibration from '../types/IImageCalibration';
@@ -63,7 +67,7 @@ class Viewport implements IViewport {
   protected flipHorizontal = false;
   protected flipVertical = false;
   public isDisabled: boolean;
-  /** Record the renddering status, mostly for testing purposes, but can also
+  /** Record the rendering status, mostly for testing purposes, but can also
    * be useful for knowing things like whether the viewport is initialized
    */
   public viewportStatus: ViewportStatus = ViewportStatus.NO_DATA;
@@ -371,6 +375,14 @@ class Viewport implements IViewport {
    */
   public getActors(): Array<ActorEntry> {
     return Array.from(this._actors.values());
+  }
+
+  /**
+   * Returns an array of unique identifiers for all the actors in the viewport.
+   * @returns An array of strings
+   */
+  public getActorUIDs(): Array<string> {
+    return Array.from(this._actors.keys());
   }
 
   /**
@@ -874,6 +886,14 @@ class Viewport implements IViewport {
       vec2.subtract(vec2.create(), initialCanvasFocal, currentCanvasFocal)
     );
     return result;
+  }
+
+  public getCurrentImageIdIndex(): number {
+    throw new Error('Not implemented');
+  }
+
+  public getTargetId(specifier?: TargetSpecifier): string {
+    return null;
   }
 
   /**