+++ packages/tools/src/stateManagement/segmentation/segmentIndex.ts
@@ -1,9 +1,15 @@
-import { getSegmentation } from './segmentationState';
+import { invalidateBrushCursor } from '../../utilities/segmentation/';
+import {
+  getSegmentation,
+  getToolGroupIdsWithSegmentation,
+} from './segmentationState';
 import { triggerSegmentationModified } from './triggerSegmentationEvents';
 
 /**
  * Set the active segment index for a segmentation Id. It fires a global state
- * modified event.
+ * modified event. Also it invalidates the brush cursor for all toolGroups that
+ * has the segmentationId as active segment (since the brush cursor color
+ * should change as well)
  *
  * @triggers SEGMENTATION_MODIFIED
  * @param segmentationId - The id of the segmentation that the segment belongs to.
@@ -20,6 +26,13 @@ function setActiveSegmentIndex(
 
     triggerSegmentationModified(segmentationId);
   }
+
+  // get all toolGroups that has the segmentationId as active
+  // segment and call invalidateBrushCursor on them
+  const toolGroups = getToolGroupIdsWithSegmentation(segmentationId);
+  toolGroups.forEach((toolGroupId) => {
+    invalidateBrushCursor(toolGroupId);
+  });
 }
 
 /**