+++ packages/tools/src/tools/index.ts
@@ -27,8 +27,11 @@ import RectangleROITool from './annotation/RectangleROITool';
 import EllipticalROITool from './annotation/EllipticalROITool';
 import CircleROITool from './annotation/CircleROITool';
 import SplineROITool from './annotation/SplineROITool';
+import SplineContourSegmentationTool from './annotation/SplineContourSegmentationTool';
 import PlanarFreehandROITool from './annotation/PlanarFreehandROITool';
+import PlanarFreehandContourSegmentationTool from './annotation/PlanarFreehandContourSegmentationTool';
 import LivewireContourTool from './annotation/LivewireContourTool';
+import LivewireContourSegmentationTool from './annotation/LivewireContourSegmentationTool';
 import ArrowAnnotateTool from './annotation/ArrowAnnotateTool';
 import AngleTool from './annotation/AngleTool';
 import CobbAngleTool from './annotation/CobbAngleTool';
@@ -47,6 +50,7 @@ import RectangleROIStartEndThresholdTool from './segmentation/RectangleROIStartE
 import BrushTool from './segmentation/BrushTool';
 import PaintFillTool from './segmentation/PaintFillTool';
 import OrientationMarkerTool from './OrientationMarkerTool';
+import SegmentSelectTool from './segmentation/SegmentSelectTool';
 
 export {
   // ~~ BASE
@@ -78,7 +82,9 @@ export {
   CircleROITool,
   SplineROITool,
   PlanarFreehandROITool,
+  PlanarFreehandContourSegmentationTool,
   LivewireContourTool,
+  LivewireContourSegmentationTool,
   ArrowAnnotateTool,
   AngleTool,
   CobbAngleTool,
@@ -92,6 +98,7 @@ export {
   SphereScissorsTool,
   RectangleROIThresholdTool,
   RectangleROIStartEndThresholdTool,
+  SplineContourSegmentationTool,
   BrushTool,
   MagnifyTool,
   AdvancedMagnifyTool,
@@ -99,4 +106,5 @@ export {
   PaintFillTool,
   ScaleOverlayTool,
   OrientationMarkerTool,
+  SegmentSelectTool,
 };